langgraph-service/requirements.txt
langgraph
langchain-core
fastapi
uvicorn
httpx
pydantic
python-dotenv
